fix: resolve all compiler warnings and macOS sed issue
- cpu_disasm.c: use sizeof(numstr) instead of BUFSTRSIZE when formatting
  into numstr[BUFSTRSIZE_SMALL] to avoid buffer size mismatch warning
- cpu_mms.c: cast ulong pageTableEntry to uint32_t for %08X format args
- nd100x.c: use %llu / unsigned long long cast for uint64_t instr_counter
- debugger.c: add missing return 0 at end of cmd_disassemble
- machine.c: handle BOOT_NONE in boot switch to silence -Wswitch
- Makefile: use SED_INPLACE variable (sed -i '' on macOS, sed -i on Linux)
  to fix cache-bust step failing on macOS with "sed: -e: No such file"

+++ src/cpu/cpu_mms.c
@@ -641,7 +641,7 @@ bool checkPageProtection(uint VPN, uint pageTable, ulong pageTableEntry, bool Us
             static int pf25 = 0;
             if (pf25 < 3)
                 printf("\r\nPF25: PTe=0x%08X pfMask=0x%08lX PIL=%d am=%d\r\n",
-                       pageTableEntry, (unsigned long)pfMask, CurrLEVEL, am);
+                       (uint32_t)pageTableEntry, (unsigned long)pfMask, CurrLEVEL, am);
             pf25++;
         }
         // PGS bit 14 (PM - Permit violation) must be FALSE here.
@@ -676,7 +676,7 @@ bool checkPageProtection(uint VPN, uint pageTable, ulong pageTableEntry, bool Us
             static int mpv25 = 0;
             if (mpv25 < 5)
                 printf("\r\nACCESS_DENIED: PT=%d VPN=%d PTe=0x%08X need=0x%08lX am=%d UseAPT=%d PIL=%d VA=%06o\r\n",
-                       pageTable, VPN, pageTableEntry, (unsigned long)accessBits, am, UseAPT, CurrLEVEL, virtualAddress);
+                       pageTable, VPN, (uint32_t)pageTableEntry, (unsigned long)accessBits, am, UseAPT, CurrLEVEL, virtualAddress);
             mpv25++;
         }
         UpdatePGS(pageTable, VPN, am, true);
